QE-1 Решить квадратное уравнение ax^2+bx+c=-7. На вход в программу подаются значения a,b,c. На выходе хотим получить решение уравнения ( в процессе )

+++ Krujka/src/main/java/com.company/Krujka.java
@@ -6,7 +6,7 @@ class Krujka {
         int a = 1;
         int b = 2;
         double c = 2.75;
-        double d = b * 2 - 4 * a * c;
+        int d = (int) (b * 2 - 4 * a * c);
         System.out.print(d);
     }
 }
